Make global effects functions compatible with minor mode functions.
Global effect functions should now activate the effect if they receive 1
as argument and deactivate them if they receive -1. This way, it is
possible to add a global minor mode to `writeroom-global-effects`
without having to write a custom function.

This implements the suggestion in #21. It also creates an
incompatibility with existing configs, so the major version number is
being bumped.